2/requirements.txt
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
customtkinter==5.2.1
darkdetect==0.8.0
idna==3.4
packaging==23.2
Pillow==9.5.0
requests==2.31.0
soupsieve==2.5
tkhtmlview==0.2.0
urllib3==2.1.0
